+++ HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ---------------
 
+1.16.2 (2021-02-05)
++++++++++++++++++++
+
+**Improvements**
+
+- Blotter strategy orders added for faster lookup
+- Strategy name hash cached
+- Minor selection_exposure optimisations
+- Simulated optimisations
+
 1.16.1 (2021-01-28)
 +++++++++++++++++++